+++ src/main/java/me/wook/springboot/data/entity/Screen.java
@@ -1,33 +1,26 @@
 package me.wook.springboot.data.entity;
 
+import lombok.*;
+
+import javax.persistence.*;
 import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-import javax.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @Builder
+@EqualsAndHashCode
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
 public class Screen {
+    @Id
+    @GeneratedValue
+    private long id;
+    @Column
+    private String name;
 
-  @Id
-  @GeneratedValue
-  private long id;
-  @Column
-  private String name;
+    @ManyToOne
+    private Movie movie;
 
-  @ManyToOne
-  private Movie movie;
-
-  @ManyToMany
-  private List<Seat> seats;
+    @ManyToMany
+    private List<Seat> seats;
 }
