+++ src/main/java/me/wook/springboot/data/entity/Seat.java
@@ -1,30 +1,24 @@
 package me.wook.springboot.data.entity;
 
+import lombok.*;
+
+import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @Builder
+@EqualsAndHashCode
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
 public class Seat {
+    @Id
+    @GeneratedValue
+    private long id;
+    @Column
+    private String name;
 
-  @Id
-  @GeneratedValue
-  private long id;
-  @Column
-  private String name;
-
-  @ManyToMany
-  private List<Screen> screens = new ArrayList<>();
+    @ManyToMany
+    private List<Screen> screens = new ArrayList<>();
 }
