+++ src/main/java/me/wook/springboot/data/repository/MovieRepository.java
@@ -6,5 +6,4 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface MovieRepository extends JpaRepository<Movie, Long> {
-
 }
