+++ src/main/java/me/wook/springboot/data/service/MovieService.java
@@ -1,17 +1,12 @@
 package me.wook.springboot.data.service;
 
 import me.wook.springboot.data.entity.Movie;
-import me.wook.springboot.web.dto.MovieDTO;
+import me.wook.springboot.web.model.MovieDTO;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface MovieService {
-
-  Page<Movie> list(final Pageable pageable) throws Exception;
-
-  Movie detail(final MovieDTO movieDTO);
-
-  boolean add(final MovieDTO movieDto);
-
-  boolean update(final MovieDTO movieDTO);
+    Page<Movie> list(final Pageable pageable) throws Exception;
+    
+    boolean add(final MovieDTO movieDto);
 }
