--- src/main/java/me/wook/springboot/movie/model/Movie.java
@@ -1,16 +1,0 @@
-package me.wook.springboot.movie.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Movie {
-
-  private long id;
-  private String name;
-}
