+++ src/main/java/me/wook/springboot/web/model/MovieDTO.java
@@ -1,0 +1,28 @@
+package me.wook.springboot.web.model;
+
+import lombok.*;
+import me.wook.springboot.data.entity.Movie;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+public class MovieDTO {
+    private long id;
+    private String name;
+    
+    public Movie movie() {
+        return Movie.builder()
+                .id(id)
+                .name(name)
+                .build();
+    }
+    
+    public Movie movie(MovieDTO movieDto) {
+        return Movie.builder()
+                .id(movieDto.getId())
+                .name(movieDto.getName())
+                .build();
+    }
+}
