+++ src/main/resources/application.yaml
@@ -6,7 +6,7 @@ spring:
     url: jdbc:mysql://127.0.0.1:13306/theater?serverTimezone=UTC&useUnicode=true&characterEncoding=utf8
     username: wook
     password: qwer1234
-  # jpa
+# jpa
   jpa:
     generate-ddl: true
     hibernate:
